Allow configurable arguments to requests.Response.json()
As with #8, this is mainly to support control over the handling of
floats; by default requests.Response.json() will parse JSON floats into
python floats, which introduces inaccuracy.

Response.json() just passes the **kwargs on to json.loads() and likewise
this change allows a dict of keywords to be set on the client.

The default behaviour is unchanged.