Support retrieving full transaction history.
When requesting full transaction history, the Oanda API only provides a
url to the location that the history will eventually be available at.
This provides a method for polling the location until it's available,
then downloading the file, processing the content of the ZIP and
extracting the result.

Note that this cannot be tested in the sandbox, as it doesn't return a
valid URL for the transaction history.